Move to activerecord dir
